@@ -18,7 +18,7 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
 
-  spec.add_runtime_dependency "profitbricks-sdk-ruby", "~> 3.0"
+  spec.add_runtime_dependency "profitbricks-sdk-ruby", "~> 4.0.0"
   spec.add_runtime_dependency "chef", "~> 12"
 
   spec.add_development_dependency "bundler", "~> 1.6"
