@@ -11,7 +11,7 @@ class Chef
              short: '-n NAME',
              long: '--name NAME',
              description: 'Name of the data center',
-             required: true
+             proc: proc { |name| Chef::Config[:knife][:name] = name }
 
       option :description,
              short: '-D DESCRIPTION',
@@ -22,25 +22,26 @@ class Chef
              short: '-l LOCATION',
              long: '--location LOCATION',
              description: 'Location of the data center',
-             proc: proc { |location| Chef::Config[:knife][:location] = location },
-             required: true
+             proc: proc { |location| Chef::Config[:knife][:location] = location }#,
 
       def run
         $stdout.sync = true
+
+        validate_required_params(%i(name location), Chef::Config[:knife])
 
         print "#{ui.color('Creating data center...', :magenta)}"
 
         connection
         datacenter = ProfitBricks::Datacenter.create(
-          name: config[:name],
-          description: config[:description],
-          location: config[:location]
+          name: Chef::Config[:knife][:name],
+          description: Chef::Config[:knife][:description],
+          location: Chef::Config[:knife][:location]
         )
 
         dot = ui.color('.', :magenta)
         datacenter.wait_for { print dot; ready? }
-        datacenter.reload
 
+        @dcid = datacenter.id
         puts "\n"
         puts "#{ui.color('ID', :cyan)}: #{datacenter.id}"
         puts "#{ui.color('Name', :cyan)}: #{datacenter.properties['name']}"
