@@ -10,24 +10,23 @@ class Chef
       option :location,
              short: '-l LOCATION',
              long: '--location LOCATION',
-             description: 'Location of the IP block (us/las, de/fra, de/fkb)',
-             required: true
+             description: 'Location of the IP block (us/las, us/ewr, de/fra, de/fkb)'
 
       option :size,
              short: '-S INT',
              long: '--size INT',
-             description: 'The number of IP addresses to reserve',
-             required: true
+             description: 'The number of IP addresses to reserve'
 
       def run
         $stdout.sync = true
+        validate_required_params(%i(size location), Chef::Config[:knife])
 
         print "#{ui.color('Allocating IP block...', :magenta)}"
 
         connection
         ipblock = ProfitBricks::IPBlock.create(
-          location: config[:location],
-          size: config[:size]
+          location: Chef::Config[:knife][:location],
+          size: Chef::Config[:knife][:size]
         )
 
         dot = ui.color('.', :magenta)
@@ -37,7 +36,7 @@ class Chef
         puts "#{ui.color('ID', :cyan)}: #{ipblock.id}"
         puts "#{ui.color('Location', :cyan)}: #{ipblock.properties['location']}"
         puts "#{ui.color('IP Addresses', :cyan)}: #{ipblock.properties['ips']}"
-
+        @ipid = ipblock.id
         puts 'done'
       end
     end
