@@ -16,11 +16,12 @@ class Chef
       option :server_id,
              short: '-S SERVER_ID',
              long: '--server-id SERVER_ID',
-             description: 'The ID of the server assigned the NIC',
-             required: true
+             description: 'The ID of the server assigned the NIC'
 
       def run
         $stdout.sync = true
+        validate_required_params(%i(datacenter_id server_id), Chef::Config[:knife])
+
         nic_list = [
           ui.color('ID', :bold),
           ui.color('Name', :bold),
@@ -31,7 +32,7 @@ class Chef
         ]
         connection
 
-        ProfitBricks::NIC.list(config[:datacenter_id], config[:server_id]).each do |nic|
+        ProfitBricks::NIC.list(Chef::Config[:knife][:datacenter_id], Chef::Config[:knife][:server_id]).each do |nic|
           nic_list << nic.id
           nic_list << nic.properties['name']
           nic_list << nic.properties['ips'].to_s
