@@ -16,14 +16,14 @@ class Chef
       option :server_id,
              short: '-S SERVER_ID',
              long: '--server-id SERVER_ID',
-             description: 'The ID of the server',
-             required: true
+             description: 'The ID of the server'
 
       def run
         connection
+        validate_required_params(%i(datacenter_id server_id), Chef::Config[:knife])
         @name_args.each do |volume_id|
           begin
-            volume = ProfitBricks::Volume.get(config[:datacenter_id], nil, volume_id)
+            volume = ProfitBricks::Volume.get(Chef::Config[:knife][:datacenter_id], nil, volume_id)
           rescue Excon::Errors::NotFound
             ui.error("Volume ID #{volume_id} not found. Skipping.")
             next
@@ -37,7 +37,7 @@ class Chef
 
           confirm('Do you really want to detach this volume')
 
-          volume.detach(config[:server_id])
+          volume.detach(Chef::Config[:knife][:server_id])
           ui.msg("Detaching volume #{volume_id} from server")
         end
       end
