@@ -4,20 +4,21 @@ require 'profitbricks_datacenter_list'
 Chef::Knife::ProfitbricksDatacenterList.load_deps
 
 describe Chef::Knife::ProfitbricksDatacenterList do
+  let(:datacenter_list) { Chef::Knife::ProfitbricksDatacenterList.new }
+
   before :each do
-    subject { Chef::Knife::ProfitbricksDatacenterList.new }
-    allow(subject).to receive(:puts)
+    allow(datacenter_list).to receive(:puts)
   end
 
   describe '#run' do
     it 'should output the column headers' do
-      expect(subject).to receive(:puts).with(/^ID\s+Name\s+Description\s+Location\s+Version\s*$/)
-      subject.run
+      expect(datacenter_list).to receive(:puts).with(/^ID\s+Name\s+Description\s+Location\s+Version\s*$/)
+      datacenter_list.run
     end
 
     it 'should output the data center locations' do
-      expect(subject).to receive(:puts).with(/(?:us\/las)/)
-      subject.run
+      expect(datacenter_list).to receive(:puts).with(/^ID\s+Name\s+Description\s+Location\s+Version\s*$/)
+      datacenter_list.run
     end
   end
 end
